#65. Delete sessions when deleting bot. Delete Images and messageLogs when deleting session
A save(session) is now performed inside the 'new' method in in lib/aida/session.ex.
This approach is taken in order to provide a valid session_id for session related entities via a belongs_to relation (i.e messageLog), so an entry with
the session_id passed as argument exists at the moment of creating the related entity. (i.e, that a session exists in the database at the moment of persisting a message log)